requirements.txt
colorama==0.4.6
jira==3.8.0
python-dotenv==1.1.0
